Cleanup main.py
This separates the remaining elements into their own router files
