+++ backend/docker-compose.yml
@@ -19,7 +19,7 @@ services:
       dockerfile: Dockerfile
       context: .
     restart: unless-stopped
-    command: bash -c "alembic upgrade head && uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"
+    command: bash -c "alembic upgrade head && uvicorn app.main:router --host 0.0.0.0 --port 8000 --reload"
     environment:
       - ENV=development
       - REDIS_HOST=redis
