+++ backend/docker-entrypoint.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 alembic upgrade head
-uvicorn app.main:app --host 0.0.0.0 --port 8000 $@
+uvicorn app.main:router --host 0.0.0.0 --port 8000 $@
